requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.2.0
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.1
mysql-connector-python==8.3.0
python-multipart

